HIVE-24204:fix
